backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyPDF2
httpx
python-dotenv
google-generativeai